Move some extension functions to new file Some documentation tweaks
